+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project are documented in this file.
 ### General
 
  - Fixed compiler panics when some complex expressions are used for the model expression in `for` (#2977)
+ - Native style: Fixed support for floating point ranges in Slider.
 
 ### Slint Language